Simplify some of the code in handleRootPolicy (#89)
* Delete unused code

The `status` variable is never nil since it's previously initialized as
an empty slice.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Simplify the code to check for stale replicated policies

This previously used a slice of decisions which lead to a time
complexity of O(n^2). This is simplified to using a set which makes the
time complexity O(n).

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Remove a check for an existing status in status slice

This seems like an impossible case since it'd require two replicated
policies to have the same root policy. Then even if that did happen,
arbitrarily picking the last one in the loop doesn't seem correct
either.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>