Retry on errors when reconciling the root policy (#90)
* Add ./go to .gitignore

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Retry on errors when reconciling the root policy

Before this, the controller would be stuck in an infinite loop in the
event something failed when reconciling the root policy. This is not
likely in most cases except for a bad template string. This enables
proper error handling of template strings rather than just ignoring
them.

This ended up refactoring a lot of the handleRootPolicy method in order
to make it easier to read with the addition of retry logic.

Resolves https://github.com/open-cluster-management/backlog/issues/15438

Signed-off-by: mprahl <mprahl@users.noreply.github.com>