Fix reconciling a policy with a dot in its name
Namespaces are required to be valid RFC 1123 DNS labels, so they do not
have a `.` in their names. But that is not required for names of other
objects, like this code might have assumed. So when a policy had a name
like `foo.bar`, an error was occurring, which prevented the root policy
from getting a status update.

Refs:
 - https://issues.redhat.com/browse/ACM-2712

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>