Refactor: create a Propagator type
The plan is to reduce work done by the propagator by considering the
types of changes in the cluster that are causing the reconcile. In order
to split the types of reconciles, new reconcilers will be necessary,
but they need to do the same kind of propagation. This new type will
help manage things those reconcilers have in common.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>