Refactor: move status reconciler into propagator
This moves the root policy status reconciler into the propagator package
in preparation for a new reconciler in that package. The alternative
would likely be to make another new package for the new reconciler...
But the new reconciler will also need some re-organization of logic
between all of the reconcilers, including some logic which needs to be
shared. So bringing it all into one package seems cleanest.

One (easy to miss) issue with the current commit is that the logger in
the root-policy-status Reconcile is now named "propagator", because it
uses the package's global log. IMO using the global log is not ideal
anyway, and I'd like to address this issue at its source somehow.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>