Use ResourceVersion to skip duplicated reconciles
This may not prevent all of the duplicated reconciles, because the
reconciler can not write the new resource version to the cache until the
API call is complete. By the time it can write the new version, the
event might have already made it to the predicate functions. Based on
metrics from the e2e tests, this may reduce the number of replicated
policy reconciles by 35%.

Events from the dependency watcher are artificially delayed on the
replicated policy reconcile queue - the watch there and a watch for the
controller-runtime cache on ManagedCluster labels will race each other,
and if the cache is old during the reconcile, the needed update will
be missed. This is not the best way to resolve the race: it would be
better to ensure that both use the same watch or cache somehow.

Refs:
 - https://issues.redhat.com/browse/ACM-7332

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>